+++ src/main/java/bowling/domain/Pins.java
@@ -1,4 +1,4 @@
-package bowling.model;
+package bowling.domain;
 
 import bowling.exception.InvalidPinsException;
