+++ src/main/java/bowling/domain/Player.java
@@ -1,4 +1,4 @@
-package bowling.model;
+package bowling.domain;
 
 import bowling.exception.InvalidUsernameException;
