+++ src/main/java/bowling/domain/frame/FrameFactory.java
@@ -1,4 +1,4 @@
-package bowling.model.frame;
+package bowling.domain.frame;
 
 import java.util.List;
 import java.util.stream.Collectors;
