+++ src/test/java/bowling/domain/frame/FrameFactoryTest.java
@@ -1,4 +1,4 @@
-package bowling.model.frame;
+package bowling.domain.frame;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
