+++ src/test/java/bowling/domain/frame/LastFrameTest.java
@@ -1,6 +1,6 @@
-package bowling.model.frame;
+package bowling.domain.frame;
 
-import bowling.model.Pins;
+import bowling.domain.Pins;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
