+++ src/test/java/bowling/domain/state/MissTest.java
@@ -1,7 +1,7 @@
-package bowling.model.state;
+package bowling.domain.state;
 
 import bowling.exception.ImpossiblePitchException;
-import bowling.model.Pins;
+import bowling.domain.Pins;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
